mcps/requirements.txt
pandas
numpy
pytest
mcp
polygon-api-client
twelvedata
requests
fastapi
uvicorn
pydantic
google-cloud-storage
google-cloud-run
python-dotenv
redis
pyarrow
google-cloud-bigquery
google-cloud-bigquery-storage
db-dtypes
pandas_gbq
